requirements.txt
Django~=4.0.6
python-dotenv~=0.20.0
djangorestframework~=3.13.1
requests~=2.27.1
setuptools~=62.6.0